% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_lambda_threshold_fast.R
\name{calculate_lambda_threshold_fast}
\alias{calculate_lambda_threshold_fast}
\title{Empirically calculate a significance threshold for Lambda.}
\usage{
calculate_lambda_threshold_fast(k, alpha)
}
\arguments{
\item{k}{Number of indices.}

\item{alpha}{Level of significance.}
}
\value{
An estimated 100(1-alpha)% quantile from the distribution of Lambda.
}
\description{
For a given significance value, this function simulates the distribution of lambda in the context of glass-fragment refractive indices and returns the quantile at
that significance level.
}
