% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_B0.R
\name{find_B0}
\alias{find_B0}
\title{Calculate B0 for a given array of Refractive Indices.}
\usage{
find_B0(array, i = 1, j = length(array))
}
\arguments{
\item{array}{vector of refractive indices.}

\item{i}{Starting element (default = 1)}

\item{j}{Ending element (default = length(array))}
}
\value{
A numeric corresponding to the maximum between-sum-of-squares estimate from the sample.
}
\description{
Calculates the B0 value for a given numeric vector of values; assuming they're appropriate values corresponding to glass fragment refractive indices.
}
