% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_T0_multi.R
\name{find_T0}
\alias{find_T0}
\title{Calculate T0 for a given list of glass fragments and features.}
\usage{
find_T0(data, i = 1, j = length(data))
}
\arguments{
\item{data}{list of glass fragment chemical (or otherwise) features.}

\item{i}{Starting element (default = 1)}

\item{j}{Ending element (default = length(array))}
}
\value{
A numeric corresponding to the maximum between-sum-of-squares estimate from the sample.
}
\description{
Calculates the "T0" value (the split corresponding to the maximum value of T^2) for a given list of data sets
corresponding to glass fragment features assuming they're appropriate values corresponding to glass fragment features.
}
