% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hotellings.R
\name{find_T2}
\alias{find_T2}
\title{Calculate Hotelling's T^2 Statistic for two independent multivariate samples.}
\usage{
find_T2(d1, d2)
}
\arguments{
\item{d1}{matrix or data.frame type object containing the multivariate data for the first sample.}

\item{d2}{matrix or data.frame type object containing the multivariate data for the second sample.}
}
\value{
T^2 value for the two objects.
}
\description{
Calculate Hotelling's T^2 Statistic for two independent multivariate samples.
}
