% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glass-data.R
\docType{data}
\name{glass}
\alias{glass}
\title{Glass Fragment Elemental Composition Data}
\format{
## `who`
}
\source{
Unknown
}
\usage{
glass
}
\description{
A dataset comprised of various glass fragment samples including multiple measurements of 7 element concentrations.
}
\keyword{datasets}
