% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glass-data.R
\docType{data}
\name{glass2}
\alias{glass2}
\title{glass2}
\format{
a `data.frame` with 16 rows and 16 columns.
\describe{
  \item{item}{factor}{761 levels - which item the measurements came from}
  \item{Li7}{numeric}{log of lithium concentration}
  \item{Mg25}{numeric}{log of magnesium concentration}
  \item{Al27}{numeric}{log of aluminium concentration}
  \item{K39}{numeric}{log of potassium concentration}
  \item{Ti49}{numeric}{log of titanium concentration}
  \item{Mn55}{numeric}{log of manganese concentration}
  \item{Fe57}{numeric}{log of iron concentration}
  \item{Rb85}{numeric}{log of rubidium concentration}
  \item{Sr88}{numeric}{log of strontium concentration}
  \item{Zr90}{numeric}{log of zirconium concentration}
  \item{Ba137}{numeric}{log of barium concentration}
  \item{La139}{numeric}{log of lanthanum concentration}
  \item{Ce140}{numeric}{log of cerium concentration}
  \item{Nd146}{numeric}{log of neodymium concentration}
  \item{Pb208}{numeric}{log of lead concentration}
}
}
\source{
Almirall, Jose; Akmeemana, Anuradha, 2022,
"casework.tab", Shiny Glass Application, https://doi.org/10.34703/gzx1-9v95/OB8BS9/CP6WXP,
FIU Research Data Portal, V2, UNF:6:jQxEQCGZVvlWtc6owbtp+A== [fileUNF]
}
\usage{
data(glass2)
}
\description{
Glass Fragment Elemental Composition Data on 15 variables.
}
\details{
Log transformed example casework data
}
\references{
Anuradha Akmeemana, R. C., Jose Almirall, The Calculation of Calibrated Likelihood Ratios (LRs) for
Glass Using a Multivariate Kernel Density Model: Introducing a User-Friendly Graphical User Interface (GUI).
In American Academy of Forensic Science, Anaheim, CA, 2020.
}
\keyword{datasets}
