% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/has_children.R
\name{has.children}
\alias{has.children}
\title{has.children}
\usage{
has.children(part)
}
\arguments{
\item{part}{Node in partition Tree.}
}
\value{
Logical determining if the node has any children.
}
\description{
For internal use only.
Determines if a node in the Partition tree has a child.
}
