% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partition.multi.R
\name{order_euclid}
\alias{order_euclid}
\title{Order a list of data frames containing numerical columns by their euclidean distance to the mean.}
\usage{
order_euclid(alist)
}
\arguments{
\item{alist}{A list of data frames.}
}
\value{
A list of data frames.
}
\description{
Meant for internal use only.
}
