% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partition_C.R
\name{partition_C}
\alias{partition_C}
\title{Create Partitions of an RI array.}
\usage{
partition_C(array, alpha = 0.05, .debug = FALSE)
}
\arguments{
\item{array}{Vector of refractive indices.}

\item{alpha}{Significance parameter "[0,1]". Higher values are more likely to partition the
array further.}

\item{.debug}{Runs debugging.}
}
\value{
None
}
\description{
Partitions the array of assumed glass fragment refractive indices into statistically
significant groups.
}
\examples{
None


}
