% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_tree.R
\name{plot.sk_partition_tree}
\alias{plot.sk_partition_tree}
\title{Plot method for objects of type "sk_partition_tree".}
\usage{
\method{plot}{sk_partition_tree}(x, ...)
}
\arguments{
\item{x}{Output from the function "partition()"}

\item{...}{Extra details for the plot. Unused.}
}
\value{
Plot of the decision tree that is formed by the Revised Scott-Knott algorithm, designed by Prof. James Curran.
}
\description{
Plot method for objects of type "sk_partition_tree".
}
\examples{

data = generate_indices()
part = partition(data)
plot.sk_partition_tree(part)
plot(part)

}
