% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_tree.R
\name{plot.tree}
\alias{plot.tree}
\title{Scott-Knott for Forensic Data Analysis (Modification 2).}
\usage{
\method{plot}{tree}(part)
}
\arguments{
\item{part}{output from the function "partition()"}
}
\value{
Plot of the decision tree that is formed by the Revised Scott-Knott algorithm, designed by Prof. James Curran.
}
\description{
SKM2. This function takes a vector of glass fragment refractive indices and returns a grouping of the values as outlined by the algorithm.
}
\examples{

data = generate_indices()
part = partition(data)
plot.tree(part)

}
