% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{prepare_data}
\alias{prepare_data}
\title{Prepare a data file that is in standard form.}
\usage{
prepare_data(data, label = NA)
}
\arguments{
\item{data}{Inputted data.frame.}

\item{label}{Column corresponding to the label wished to be grouped by.}
}
\value{
A list of split data.
}
\description{
Expects that the first two columns correspond to the item and fragment information, and that the remaining columns are the features.
}
