% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hotellings.R
\name{ptsquared}
\alias{ptsquared}
\title{Calculate the Probability for a given T^2 statistic.}
\usage{
ptsquared(t, n1, n2, p)
}
\arguments{
\item{t}{T^2 statistic.}

\item{n1}{Number of observations in first sample.}

\item{n2}{Number of observations in second sample.}

\item{p}{Number of parameters.}
}
\value{
A probability corresponding to a given T^2 statistic and for given arguments.
}
\description{
Calculate the Probability for a given T^2 statistic.
}
