% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partition.R
\name{ungroup.partition}
\alias{ungroup.partition}
\title{ungroup.partition}
\usage{
ungroup.partition(tree)
}
\arguments{
\item{tree}{tree object returned from partition()}
}
\value{
A list object containing the indices of the
}
\description{
Ungroups the tree object in the output from partition()
}
