% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_lambda_threshold.R
\name{calculate_lambda_threshold}
\alias{calculate_lambda_threshold}
\title{Empirically calculate a significance threshold for Lambda.}
\usage{
calculate_lambda_threshold(k, alpha, .nsims = 10000)
}
\arguments{
\item{k}{Number of indices.}

\item{alpha}{Level of significance.}

\item{.nsims}{Number of simulated arrays. More simulations implies a closer estimated distribution of the true lambda distribution.}
}
\value{
An estimated 100(1-alpha)% quantile from the distribution of Lambda.
}
\description{
For a given significance value, this function simulates the distribution of lambda in the context of glass-fragment refractive indices and returns the quantile at
that significance level.
}
