% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_lambda_threshold.R
\name{calculate_lambda_threshold}
\alias{calculate_lambda_threshold}
\title{Calculate the significance threshold for Lambda.}
\usage{
calculate_lambda_threshold(k, alpha)
}
\arguments{
\item{k}{Number of indices.}

\item{alpha}{Level of significance.}
}
\value{
A 100(1-alpha)\% quantile estimate from the distribution of Lambda.
}
\description{
For a given significance value, this function uses critical values determined from simulated data formed on 1 million arrays, and returns the quantile estimated at
that significance level.
For values of k > 20, it assumes a chi squared distribution with k/(pi - 2) degrees of freedom.
}
