% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{find_B0}
\alias{find_B0}
\title{Calculate B0}
\usage{
find_B0(arr)
}
\arguments{
\item{arr}{vector of refractive indices.}
}
\value{
A numeric corresponding to the maximum between-sum-of-squares estimate from the sample.
}
\description{
Calculates the B0 value for a given numeric vector of values; assuming they're appropriate values corresponding to glass fragment refractive indices.
}
