% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_tree.R
\name{plot.sk_partition_tree}
\alias{plot.sk_partition_tree}
\title{Plot S3method for objects of type "sk_partition_tree".}
\usage{
\method{plot}{sk_partition_tree}(x, ...)
}
\arguments{
\item{x}{Output from the function "partition()"}

\item{...}{Extra details for the plot. Unused.}
}
\value{
Plot of the decision tree that is formed by the sk_partition_tree object returned by partition and partition.multi.
}
\description{
S3method for plotting the resulting tree formed by the partitioning algorithms in the SK4FGA package.
}
\examples{

data = generate_indices()
part = partition(data)
plot.sk_partition_tree(part)

data(glass)
data.multi = prepare_data(glass, 1)[1:3]
part = partition.multi(data.multi)
plot(part)

}
