% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{prepare_data}
\alias{prepare_data}
\title{prepare_data}
\usage{
prepare_data(data, label = NA)
}
\arguments{
\item{data}{Inputted data.frame.}

\item{label}{Column corresponding to the label to be grouped by.}
}
\value{
A list of split data.
}
\description{
Prepare a data file that is in standard form for partition.multi.
}
